FU-84 feat: add middleware for handle redirect
