--- src/app/(main)/_login/page.tsx
@@ -1,13 +1,0 @@
-import LoginButton from "@/components/buttons/login-button";
-
-const Page = () => {
-  return (
-    <div>
-      <h3>사진작가의 든든한 업무 파트너</h3>
-      <h3>프리비와 함께 더 자유롭게 촬영하세요</h3>
-      <LoginButton />
-    </div>
-  );
-};
-
-export default Page;
